pyproject.toml
[project]
name = "spectral-lib-cleaning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "downloaders>=1.0.22",
    "matchms>=0.31.0",
    "python-dotenv>=1.2.2",
    "tqdm>=4.67.2",
]

[dependency-groups]
dev = ["jupyterlab>=4.5.3"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
spectral-clean = "spectral_lib_cleaning.cli:main"
